requirements.txt
openai
pydantic
python-dotenv
typing-extensions
